app/requirements.txt
requests==2.22.0
pyquery==1.4.1
click8==8.0.1
tldextract==2.2.2
